Convert `EmailHelper` to `CanonicalEmail` model concern
