+++ app/models/canonical_email_block.rb
@@ -12,7 +12,7 @@
 #
 
 class CanonicalEmailBlock < ApplicationRecord
-  include EmailHelper
+  include CanonicalEmail
   include Paginable
 
   belongs_to :reference_account, class_name: 'Account', optional: true
@@ -26,7 +26,7 @@ class CanonicalEmailBlock < ApplicationRecord
   end
 
   def email=(email)
-    self.canonical_email_hash = email_to_canonical_email_hash(email)
+    self.canonical_email_hash = self.class.email_to_canonical_email_hash(email)
   end
 
   def self.block?(email)
